+++ service-implementation/src/main/java/com/liaison/mailbox/service/core/processor/FileWriter.java
@@ -18,8 +18,10 @@ import com.liaison.mailbox.enums.EntityStatus;
 import com.liaison.mailbox.enums.ExecutionState;
 import com.liaison.mailbox.enums.Messages;
 import com.liaison.mailbox.enums.ProcessorType;
+import com.liaison.mailbox.rtdm.dao.StagedFileDAO;
 import com.liaison.mailbox.rtdm.dao.StagedFileDAOBase;
 import com.liaison.mailbox.rtdm.model.StagedFile;
+import com.liaison.mailbox.service.dto.configuration.processor.properties.FileWriterPropertiesDTO;
 import com.liaison.mailbox.service.exception.MailBoxConfigurationServicesException;
 import com.liaison.mailbox.service.exception.MailBoxServicesException;
 import com.liaison.mailbox.service.glass.util.GlassMessage;
@@ -41,6 +43,9 @@ import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
+import java.nio.file.Files;
+import java.nio.file.StandardCopyOption;
+import java.sql.Timestamp;
 import java.util.List;
 
 /**
@@ -227,8 +232,40 @@ public class FileWriter extends AbstractProcessor implements MailBoxProcessorI {
                         }
                     }
                 }
+                
+                persistTriggerFileEntry(workTicket, processorPayloadLocation);
             }
         }
+    }
+
+    /**
+     * To persist trigger file entry in staged file.
+     * For Shell MFT changes : GMB-1100
+     * 
+     * @param workTicket
+     * @param processorPayloadLocation 
+     */
+    private void persistTriggerFileEntry(WorkTicket workTicket, String processorPayloadLocation) {
+
+        StagedFile stagedFile = new StagedFile();
+        stagedFile.setPguid(MailBoxUtil.getGUID());
+        Timestamp timeStamp = MailBoxUtil.getTimestamp();
+        stagedFile.setCreatedDate(timeStamp);
+        stagedFile.setModifiedDate(timeStamp);
+        stagedFile.setProcessorId(configurationInstance.getPguid());
+        stagedFile.setGlobalProcessId(workTicket.getAdditionalContext().get(MailBoxConstants.KEY_TRIGGER_FILE_PARENT_GPID).toString());
+        stagedFile.setStagedFileStatus(EntityStatus.ACTIVE.name());
+        stagedFile.setFilePath(processorPayloadLocation);
+        stagedFile.setFileName(workTicket.getAdditionalContext().get(MailBoxConstants.KEY_TRIGGER_FILE_NAME).toString());
+        stagedFile.setSpectrumUri(workTicket.getAdditionalContext().get(MailBoxConstants.KEY_TRIGGER_FILE_URI).toString());
+        stagedFile.setClusterType(MailBoxUtil.CLUSTER_TYPE);
+        stagedFile.setProcessDc(MailBoxUtil.DATACENTER_NAME);
+        stagedFile.setFileSize("0");
+        stagedFile.setMailboxId((null != workTicket.getAdditionalContext().get(MailBoxConstants.KEY_MAILBOX_ID))
+                ? workTicket.getAdditionalContext().get(MailBoxConstants.KEY_MAILBOX_ID).toString() : null);
+        
+        StagedFileDAO stagedFileDAO = new StagedFileDAOBase();
+        stagedFileDAO.persist(stagedFile);
     }
 
     /**
@@ -359,9 +396,23 @@ public class FileWriter extends AbstractProcessor implements MailBoxProcessorI {
         //write the file
         FileOutputStream outputStream = null;
         try {
-            outputStream = new FileOutputStream(file);
+
+            //add status indicator if specified to indicate that uploading is in progress
+            String statusIndicator = getStatusIndicator();
+            String stagingFileName = (MailBoxUtil.isEmpty(statusIndicator))
+                    ? file.getAbsolutePath()
+                    : file.getAbsolutePath() + MailBoxConstants.DOT_OPERATOR + statusIndicator;
+            File stagingFile = new File(stagingFileName);
+
+            outputStream = new FileOutputStream(stagingFile);
             long fileSize = (long) IOUtils.copy(response, outputStream);
             workTicket.setPayloadSize(fileSize);
+            
+            // Renames the uploaded file to original extension if the fileStatusIndicator is given by User
+            if (!MailBoxUtil.isEmpty(statusIndicator)) {
+                Files.move(stagingFile.toPath(), file.toPath(), StandardCopyOption.REPLACE_EXISTING);
+            }
+
         } catch (Exception e) {
             throw new RuntimeException(e);
         } finally {
@@ -379,6 +430,22 @@ public class FileWriter extends AbstractProcessor implements MailBoxProcessorI {
                 configurationInstance.getProcessorType().name(),
                 this.isDirectUploadEnabled());
         
+    }
+
+    /**
+     * It returns the file transfer status indicator string.
+     * It only applicable for filewriter.
+     * 
+     * @return
+     * @throws IOException
+     * @throws IllegalAccessException
+     */
+    private String getStatusIndicator() throws IOException, IllegalAccessException {
+
+        if (ProcessorType.FILEWRITER.equals(configurationInstance.getProcessorType())) {
+            return ((FileWriterPropertiesDTO) getProperties()).getFileTransferStatusIndicator();
+        }
+        return null;
     }
 
     /**