+++ service-implementation/src/main/java/com/liaison/mailbox/service/dto/configuration/processor/properties/FTPUploaderPropertiesDTO.java
@@ -50,6 +50,7 @@ public class FTPUploaderPropertiesDTO extends StaticProcessorPropertiesDTO {
     private int scriptExecutionTimeout;
     private int staleFileTTL;
     private boolean useFileSystem = true;
+    private String triggerFile;
 
 	public String getUrl() {
 		return url;
@@ -193,5 +194,11 @@ public class FTPUploaderPropertiesDTO extends StaticProcessorPropertiesDTO {
     }
     public void setUseFileSystem(boolean useFileSystem) {
         this.useFileSystem = useFileSystem;
+    }
+    public String getTriggerFile() {
+        return triggerFile;
+    }
+    public void setTriggerFile(String triggerFile) {
+        this.triggerFile = triggerFile;
     }
 }